backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-cloud-pubsub==2.18.4
google-cloud-vision==3.5.0
google-cloud-firestore==2.13.1
pydantic==2.6.1
python-dotenv==1.0.1
firebase-admin==6.4.0
python-multipart==0.0.9 